+++ packages/terra-date-input/src/DateInputUtil.js
@@ -229,6 +229,15 @@ class DateInputUtil {
 
     return 'day-month-year';
   }
+
+  static getDateFormat(props) {
+    return (
+      (
+        DateInputUtil.computedDisplayFormat(props.displayFormat, props.intl.locale) === 'month-day-year')
+        ? `(${props.intl.formatMessage({ id: 'Terra.date.input.monthLabel' })} ${props.intl.formatMessage({ id: 'Terra.date.input.dayFormatLabel' })} ${props.intl.formatMessage({ id: 'Terra.date.input.yearFormatLabel' })})`
+        : `(${props.intl.formatMessage({ id: 'Terra.date.input.dayFormatLabel' })} ${props.intl.formatMessage({ id: 'Terra.date.input.monthLabel' })} ${props.intl.formatMessage({ id: 'Terra.date.input.yearFormatLabel' })})`
+    );
+  }
 }
 
 DateInputUtil.inputType = {