+++ packages/terra-date-picker/src/react-datepicker/index.jsx
@@ -195,9 +195,13 @@ class DatePicker extends React.Component {
      */
     onWeekSelect: PropTypes.func,
     /**
-     * A callback function to execute when the mouse cursor leaves the element.
+     * **Deprecated**, A callback function to execute when the mouse cursor leaves the element. Resolves to `onRequestClose`.
      */
     onClickOutside: PropTypes.func,
+    /**
+   * A callback function to execute when picker is dismissed. onRequestClose(event)
+     */
+    onRequestClose: PropTypes.func,
     /**
      * A callback function to execute when date is entered.
      */
@@ -222,10 +226,6 @@ class DatePicker extends React.Component {
      * A prop to check next month dates.
      */
     peekNextMonth: PropTypes.bool,
-    /**
-     * Value for placeholder of date picker.
-     */
-    placeholderText: PropTypes.string,
     /**
      * Prop to prevent date picker from opening on focus.
      */
@@ -393,7 +393,7 @@ class DatePicker extends React.Component {
       // If date picker is open in overlay
       if (this.datePickerOverlayContainer.current) {
         if (event.target === this.datePickerOverlayContainer.current || this.datePickerOverlayContainer.current.contains(event.target)) {
-          this.setOpen(false);
+          this.setOpen(false, event);
         }
       }
     }
@@ -424,7 +424,7 @@ class DatePicker extends React.Component {
 
   handleOnRequestClose() {
     this.setState({ isCalendarKeyboardFocused: false, isCalendarOpenedViaKeyboard: false });
-    this.setOpen(false);
+    this.setOpen(false, event);
   }
 
   getPreSelection = () => (
@@ -460,17 +460,27 @@ class DatePicker extends React.Component {
   }
 
   setOpen = (open) => {
+    if(!open) {
+      this.setState({
+        isCalendarOpenedViaKeyboard: false,
+        isCalendarKeyboardFocused: false
+      })
+    }
     this.setState({
       open: open,
       preSelection: open && this.state.open ? this.state.preSelection : this.calcInitialState().preSelection
     })
+
+    if (this.props.onRequestClose && (!open)) {
+      this.props.onRequestClose(event);
+    }
   }
 
   handleFocus = (event) => {
     if (!this.state.preventFocus) {
       this.props.onFocus(event)
       if (!this.props.preventOpenOnFocus) {
-        this.setOpen(true)
+        this.setOpen(true, event)
       }
     }
   }
@@ -499,7 +509,7 @@ class DatePicker extends React.Component {
 
   handleCalendarClickOutside = (event) => {
     if (!this.props.inline) {
-      this.setOpen(false)
+      this.setOpen(false, event)
     }
     this.props.onClickOutside(event)
     if (this.props.withPortal) { event.preventDefault() }
@@ -532,7 +542,7 @@ class DatePicker extends React.Component {
     if (!this.props.shouldCloseOnSelect) {
       this.setPreSelection(date)
     } else if (!this.props.inline) {
-      this.setOpen(false)
+      this.setOpen(false, event)
     }
   }
 
@@ -586,7 +596,7 @@ class DatePicker extends React.Component {
   }
 
   onInputKeyDown = (event) => {
-    if(event.key === 'Enter') {
+    if(event.keyCode === KeyCode.KEY_RETURN || event.keyCode === KeyCode.KEY_SPACE) {
       this.setState({ isCalendarOpenedViaKeyboard: true })
 
     }
@@ -613,11 +623,11 @@ class DatePicker extends React.Component {
         this.handleSelect(copy, event)
         !this.props.shouldCloseOnSelect && this.setPreSelection(copy)
       } else {
-        this.setOpen(false)
+        this.setOpen(false, event)
       }
     } else if (eventKey === 'Escape') {
       event.preventDefault()
-      this.setOpen(false)
+      this.setOpen(false, event)
     } else if (!this.props.disabledKeyboardNavigation && keyboardNavKeys.indexOf(eventKey) !== -1) {
       let newSelection
       switch (eventKey) {
@@ -814,7 +824,6 @@ class DatePicker extends React.Component {
       id: this.props.id,
       name: this.props.name,
       autoFocus: this.props.autoFocus,
-      placeholder: this.props.placeholderText,
       disabled: this.props.disabled,
       autoComplete: this.props.autoComplete,
       className: classNameList,