+++ packages/terra-time-input/src/TimeInput.jsx
@@ -6,6 +6,7 @@ import ThemeContext from 'terra-theme-context';
 import Input from 'terra-form-input';
 import ButtonGroup from 'terra-button-group';
 import { injectIntl, intlShape } from 'react-intl';
+import uuidv4 from 'uuid/v4';
 
 import * as KeyCode from 'keycode-js';
 import TimeUtil from './TimeUtil';
@@ -88,7 +89,7 @@ const propTypes = {
   value: PropTypes.string,
   /**
    * Type of time input to initialize. Must be `24-hour` or `12-hour`.
-   * The `de`, `es-ES`, `fr-FR`, `fr`, `nl-BE`, `nl`, `pt-BR`, `pt`, `sv-SE` and `sv` locales do not use the 12-hour time notation.
+   * The `de`, `es-ES`, `es`, `fr-FR`, `fr`, `nl-BE`, `nl`, `pt-BR`, `pt`, `sv-SE` and `sv` locales do not use the 12-hour time notation.
    * If the `variant` prop if set to `12-hour` for one of these supported locales, the variant will be ignored and defaults to `24-hour`.
    */
   variant: PropTypes.oneOf([TimeUtil.FORMAT_12_HOUR, TimeUtil.FORMAT_24_HOUR]),
@@ -133,6 +134,8 @@ class TimeInput extends React.Component {
       value = undefined;
     }
 
+    this.uuid = uuidv4();
+
     this.timeInputContainer = React.createRef();
     this.handleHourChange = this.handleHourChange.bind(this);
     this.handleMinuteChange = this.handleMinuteChange.bind(this);
@@ -155,18 +158,8 @@ class TimeInput extends React.Component {
     let meridiem;
 
     if (TimeUtil.getVariantFromLocale(props) === TimeUtil.FORMAT_12_HOUR) {
-      if (!this.props.intl.messages['Terra.timeInput.am'] || !this.props.intl.messages['Terra.timeInput.pm']) {
-        if (process.env !== 'production') {
-          // eslint-disable-next-line no-console
-          console.warn('This locale only uses 24 hour clock. The ante meridiem and post meridiem will not be displayed');
-        }
-
-        this.anteMeridiem = '';
-        this.postMeridiem = '';
-      } else {
-        this.anteMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.am' });
-        this.postMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.pm' });
-      }
+      this.anteMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.am' });
+      this.postMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.pm' });
 
       if (hour) {
         const parsedHour = TimeUtil.parseTwelveHourTime(hour, this.anteMeridiem, this.postMeridiem);
@@ -175,6 +168,14 @@ class TimeInput extends React.Component {
       } else {
         meridiem = this.anteMeridiem;
       }
+    }
+    if (this.props.variant === TimeUtil.FORMAT_12_HOUR && TimeUtil.getVariantFromLocale(props) === TimeUtil.FORMAT_24_HOUR) {
+      if (process.env !== 'production') {
+        // eslint-disable-next-line no-console
+        console.warn('This locale only uses 24 hour clock. The ante meridiem and post meridiem will not be displayed');
+      }
+      this.anteMeridiem = '';
+      this.postMeridiem = '';
     }
 
     this.state = {
@@ -703,7 +704,6 @@ class TimeInput extends React.Component {
       { 'is-focused': this.state.isFocused },
       { 'is-invalid': isInvalid },
       { 'is-incomplete': (isIncomplete && required && !isInvalid && !isInvalidMeridiem) },
-      theme.className,
     ),
     customProps.className);
 
@@ -724,11 +724,17 @@ class TimeInput extends React.Component {
       { 'initial-focus': this.state.secondInitialFocused },
     ]);
 
+    const formatDescriptionId = `terra-time-input-description-format-${this.uuid}`;
+
+    const format = showSeconds
+      ? `(${intl.formatMessage({ id: 'Terra.timeInput.hh' })}:${intl.formatMessage({ id: 'Terra.timeInput.mm' })}:${intl.formatMessage({ id: 'Terra.timeInput.ss' })})`
+      : `(${intl.formatMessage({ id: 'Terra.timeInput.hh' })}:${intl.formatMessage({ id: 'Terra.timeInput.mm' })})`;
+
     return (
       <div
         {...customProps}
         ref={this.timeInputContainer}
-        className={cx('time-input-container')}
+        className={cx('time-input-container', theme.className)}
       >
         <div className={timeInputClassNames}>
           <input
@@ -750,7 +756,6 @@ class TimeInput extends React.Component {
             type="text"
             value={this.state.hour}
             name={'terra-time-hour-'.concat(name)}
-            placeholder={intl.formatMessage({ id: 'Terra.timeInput.hh' })}
             maxLength="2"
             onChange={this.handleHourChange}
             onKeyDown={this.handleHourInputKeyDown}
@@ -759,6 +764,7 @@ class TimeInput extends React.Component {
             size="2"
             pattern="\d*"
             disabled={disabled}
+            aria-describedby={formatDescriptionId}
           />
           <span className={cx('time-spacer')}>:</span>
           <Input
@@ -770,7 +776,6 @@ class TimeInput extends React.Component {
             type="text"
             value={this.state.minute}
             name={'terra-time-minute-'.concat(name)}
-            placeholder={intl.formatMessage({ id: 'Terra.timeInput.mm' })}
             maxLength="2"
             onChange={this.handleMinuteChange}
             onKeyDown={this.handleMinuteInputKeyDown}
@@ -779,6 +784,7 @@ class TimeInput extends React.Component {
             size="2"
             pattern="\d*"
             disabled={disabled}
+            aria-describedby={formatDescriptionId}
           />
           {showSeconds && (
             <React.Fragment>
@@ -792,7 +798,6 @@ class TimeInput extends React.Component {
                 type="text"
                 value={this.state.second}
                 name={'terra-time-second-'.concat(name)}
-                placeholder={intl.formatMessage({ id: 'Terra.timeInput.ss' })}
                 maxLength="2"
                 onChange={this.handleSecondChange}
                 onKeyDown={this.handleSecondInputKeyDown}
@@ -801,6 +806,7 @@ class TimeInput extends React.Component {
                 size="2"
                 pattern="\d*"
                 disabled={disabled}
+                aria-describedby={formatDescriptionId}
               />
             </React.Fragment>
           )}
@@ -825,6 +831,9 @@ class TimeInput extends React.Component {
             />
           </ButtonGroup>
         )}
+        <div id={formatDescriptionId} className={cx('format-text')} aria-label={`${intl.formatMessage({ id: 'Terra.timeInput.timeFormatLabel' })} ${format}`}>
+          {format}
+        </div>
       </div>
     );
     /* eslint-enable jsx-a11y/no-static-element-interactions */