+++ packages/terra-time-input/tests/jest/TimeInput.test.jsx
@@ -12,6 +12,8 @@ const mockEvent = {
   preventDefault: jest.fn(),
 };
 
+jest.mock('uuid/v4', () => () => '00000000-0000-0000-0000-000000000000');
+
 beforeEach(() => {
   mockEvent.preventDefault.mockClear();
 });
@@ -324,7 +326,7 @@ it('should validate the getVariantFromLocale helper method', () => {
   const wrapper = mountWithIntl(timeInput);
   const testComponent = wrapper.children();
 
-  if (['ar', 'en', 'en-AU', 'en-CA', 'en-GB', 'en-US', 'es', 'es-US'].includes(testComponent.instance().props.intl.locale)) {
+  if (['ar', 'en', 'en-AU', 'en-CA', 'en-GB', 'en-US', 'es-US'].includes(testComponent.instance().props.intl.locale)) {
     expect(TimeUtil.getVariantFromLocale(testComponent.instance().props)).toEqual('12-hour');
   } else {
     expect(TimeUtil.getVariantFromLocale(testComponent.instance().props)).toEqual('24-hour');