+++ packages/terra-application-name/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.33.0 - (August 18, 2020)
+
+* Added
+  * Added missing themeable variables.
+  
 ## 3.32.0 - (August 4, 2020)
 
 * Changed