+++ packages/terra-application-utility/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Unreleased
 
+## 2.38.0 - (August 18, 2020)
+
 * Added
-  * Added an interaction media query to apply `hover` styles only on non-touch devices and removed hover wdio 
+  * Added an interaction media query to apply `hover` styles only on non-touch devices and removed hover wdio
+  * Added missing themeable variables.
 
 * Changed
   * Updated Changelog format