+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## Unreleased
 
+## 4.47.0 - (September 16, 2020)
+
+* Changed
+  * `onClickOutside` prop has been deprecated, will resolve to `onRequestClose`
+  * Updated wdio snapshots due to changes in date-picker and time-input
+
+* Changed
+  * Update `timeVariant` prop description to include `es` in locales supporting only 24 hour format.
+
+## 4.46.0 - (September 10, 2020)
+
+* Fixed
+  * Fix to clear out invalid date value when time input selected in date-time-picker.
+
+## 4.45.0 - (August 18, 2020)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.44.0 - (August 11, 2020)
 
 * Changed