+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## Unreleased
 
+## 3.48.0 - (September 16, 2020)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.47.0 - (September 10, 2020)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.46.0 - (August 27, 2020)
+
+* Changed
+  * Bump terra-notification-dialog version from ^3 to ^4
+
+## 3.45.0 - (August 18, 2020)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.44.0 - (August 11, 2020)
 
 * Changed
@@ -415,5 +435,4 @@
 
 ## 1.0.0 - (June 29, 2018)
 
-* Added
 * Initial stable release