+++ packages/terra-time-input/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+## 4.32.0 - (September 16, 2020)
+
+* Removed
+  * Removed `placeholder` from the input.
+
+* Added
+  * Added format to be displayed under the time-input.
+
+* Fixed
+  * Fix to allow only 24 hour format for `es` locale.
+
 ## 4.31.0 - (August 4, 2020)
 
 * Changed